openpangu: narrow SWA reuse-key fields to 32-bit
The openpangu_swa_window_view reuse key stored n_kv/n_tokens/window/pad as
int64_t, but these are bounded well under 2^31 (window/pad are uint32_t at
source; n_kv/n_tokens <= context length). Narrow to int32_t/uint32_t and drop
the widening casts. w_view/win_off stay int64_t: they feed ggml view
dims/offsets. Per review.

+++ src/llama-context.h
@@ -421,10 +421,10 @@ struct llama_context {
 
     struct openpangu_swa_window_view_state {
         bool active       = false;
-        int64_t n_kv      = 0;
-        int64_t n_tokens  = 0;
-        int64_t window    = 0;
-        int64_t pad       = 0;
+        int32_t n_kv      = 0;
+        int32_t n_tokens  = 0;
+        uint32_t window   = 0;
+        uint32_t pad      = 0;
         int64_t w_view    = 0;
         int64_t win_off   = 0;
     } openpangu_swa_window_view;
